app/requirements.txt
altgraph==0.17.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
ConfigUpdater==3.1.1
cryptography==41.0.2
et-xmlfile==1.1.0
fillpdf==0.7.2
idna==3.4
msal==1.23.0
packaging==23.1
pdf2image==1.16.3
pdfrw2==0.5.0
pefile==2023.2.7
Pillow==10.0.0
pip-review==1.3.0
pycparser==2.21
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
PyJWT==2.8.0
PyMuPDF==1.22.5
pystray==0.19.4
pywin32==306
pywin32-ctypes==0.2.2
requests==2.31.0
six==1.16.0
urllib3==2.0.4
